extract TheKeySession
