+++ gto-support-api-base/src/main/java/org/ccci/gto/android/common/api/TheKeySession.java
@@ -1,0 +1,50 @@
+package org.ccci.gto.android.common.api;
+
+import android.content.SharedPreferences;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.text.TextUtils;
+
+import java.util.Locale;
+
+public class TheKeySession extends Session {
+    @Nullable
+    private final String mGuid;
+
+    public TheKeySession(@Nullable final String id, @Nullable final String guid) {
+        this(id, guid, PREF_SESSION_BASE_NAME);
+    }
+
+    public TheKeySession(@Nullable final String id, @Nullable final String guid, @NonNull final String baseAttrName) {
+        super(id, (guid != null ? guid.toUpperCase(Locale.US) + "." : "") + baseAttrName);
+        mGuid = guid != null ? guid.toUpperCase(Locale.US) : null;
+    }
+
+    public TheKeySession(@NonNull final SharedPreferences prefs, @Nullable final String guid) {
+        this(prefs, guid, PREF_SESSION_BASE_NAME);
+    }
+
+    public TheKeySession(@NonNull final SharedPreferences prefs, @Nullable final String guid,
+                         @NonNull final String baseAttrName) {
+        super(prefs, (guid != null ? guid.toUpperCase(Locale.US) + "." : "") + baseAttrName);
+        mGuid = guid != null ? guid.toUpperCase(Locale.US) : null;
+    }
+
+    @Override
+    public boolean isValid() {
+        return super.isValid() && mGuid != null;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        final TheKeySession that = (TheKeySession) o;
+        return super.equals(o) && TextUtils.equals(mGuid, that.mGuid);
+    }
+}
